Move the value transfer into the event handlers
Before this commit, the work of moving the interim value out of the
`focus_field` and into the `search_field` and then clearing the
`focus_field` was split into separate places:

- `get_search_field_value` was doing the copying.
- and then we had to remember to clear the `focus_field`, if it existed, anytime we cleared the `search_field`

This commit moves all of that into one place, the event handlers where
the `focus_field` passes control back to the rest of Chosen.

Note: the `setTimeout`s are needed for the cut/paste events to make sure the
value being read from the `focus_field` happens after the events finish
changing the input's value.